% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bg1}
\alias{bg1}
\title{Dataframe of background cells 2000*2000.}
\format{
A data frame with 4915 rows and 2 variables:
\describe{
\item{Cell.X.Position}{X coordinate}
\item{Cell.Y.Position}{Y coordinate}
\item{Phenotype}{cell type}
...
}
}
\usage{
bg1
}
\description{
A data frame containing 4915 background cells on 2000 by 2000 size window
}
\keyword{datasets}
