% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sce.R
\name{format_sce}
\alias{format_sce}
\title{format_sce}
\usage{
format_sce(data)
}
\arguments{
\item{data}{A data frame where each row contains information about a cell.
The columns of the data frame will become the colData of the
SingleCellExperiment object.}
}
\value{
An SingleCellExperiment object
}
\description{
Format a data frame object into a SingleCellExperiment class
object where the count assay is empty.
}
