% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_spe}
\alias{format_spe}
\title{format_spe}
\usage{
format_spe(df)
}
\arguments{
\item{df}{A data frame where each row contains information about a cell.
The columns of the data frame will become the colData of the
\code{SpatialExperiment} object.}
}
\value{
An SPE object
}
\description{
Format a data frame object into a \code{SpatialExperiment} class
object where the count assay is empty.
}
