% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_images_with_clusters.R
\name{multiple_images_with_clusters}
\alias{multiple_images_with_clusters}
\title{multiple_images_with_clusters}
\usage{
multiple_images_with_clusters(
  background_sample = bg1,
  cluster_shape = 2,
  infiltration = 0.1,
  cluster_size = seq(200, 1000, 100),
  cluster_loc_x = 0,
  cluster_loc_y = 0,
  plot.image = TRUE
)
}
\arguments{
\item{plot.image}{}
}
\value{

}
\description{
multiple_images_with_clusters
}
