% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_images_with_immune_rings.R
\name{multiple_images_with_immune_rings}
\alias{multiple_images_with_immune_rings}
\title{multiple_images_with_immune_rings}
\usage{
multiple_images_with_immune_rings(
  background_sample = bg1,
  cluster_size = 200,
  ring_shape = 1,
  infiltration = 0,
  ring_width = seq(50, 100, 10),
  cluster_loc_x = 0,
  cluster_loc_y = 0,
  ring_infiltration = seq(0, 0.2, 0.05),
  plot.image = TRUE
)
}
\arguments{
\item{background_sample}{Data.frame or SingleCellExperiment object with locations
of points representing background cells. Further cell types will be simulated
based on this background sample.}

\item{cluster_size}{Numeric or Vector. The size of the cluster. If numeric,
all simulated images have the same cluster size. If vector, images with
a range of different cluster sizes will be simulated.}

\item{ring_shape}{List. Properties of immune rings to simulate. Either choose
one of the pre-designed shapes (1, 2, or 3), or manually
define the properties -- format:
R_shape3 = list(I1 = list(name_of_cluster_cell = "Tumour",size = 100,
shape = "Circle",centre_loc = data.frame("x" = 1000, "y" = 800),
infiltration_types = c("Immune", "Others"), infiltration_proportions = c(0, 0.1),
name_of_ring_cell = "Immune", immune_ring_width = 50, immune_ring_infiltration_types
= c("Tumour", "Others"), immune_ring_infiltration_proportions = c(0.3, 0.15)),
I2 = list(name_of_cluster_cell = "Tumour",size = 50,shape = "Oval",centre_loc =
data.frame("x" = 1100, "y" = 1050), infiltration_types = c("Immune", "Others"),
infiltration_proportions = c(0, 0.1), name_of_ring_cell = "Immune",
immune_ring_width = 50, immune_ring_infiltration_types = c("Tumour","Others"),
immune_ring_infiltration_proportions = c(0.3, 0.15)))}

\item{infiltration}{Numeric or Vector. The degree of infiltration.}

\item{ring_width}{Numeric or Vector. The width of the immune ring.}

\item{cluster_loc_x}{Numeric or Vector. The X location of the cluster center.
If numeric, all simulated images have the same center X location. If vector,
images with a range of different center locations will be simulated.}

\item{cluster_loc_y}{Numeric or Vector of the same length of \code{cluster_loc_x}.
The Y location of the cluster center.}

\item{ring_infiltration}{Numeric or Vector. The degree of tumour infiltration
in the region of immune rings. (Only applied to cell type "Tumour")}

\item{plot.image}{Boolean Whether plot the simulated images or not.Default is TRUE.}
}
\value{
A list of sce objects
}
\description{
multiple_images_with_immune_rings
}
\examples{
set.seed(610)
ring_image_list <- multiple_images_with_immune_rings(background_sample = bg1,
ring_shape = 1, infiltration = 0, ring_width = seq(50,100,10),
cluster_size = 300, cluster_loc_x = 0, cluster_loc_y = 0,
ring_infiltration = seq(0, 0.2,0.05), plot.image = TRUE)
}
