% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells.R
\name{plot_cells}
\alias{plot_cells}
\title{plot_cells}
\usage{
plot_cells(
  sce_object,
  categories_of_interest = NULL,
  colour_vector = NULL,
  feature_colname = "Cell.Type"
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object or data.frame that has cell locations and cell type info}

\item{categories_of_interest}{Vector of cell categories to be coloured}

\item{colour_vector}{Vector specifying the colours of each cell phenotype}

\item{feature_colname}{String specifying the column the cell categories belong to}
}
\value{
A plot is returned
}
\description{
Produces a scatter plot of the cells in the tissue. Cells are coloured
categorically by phenotype. Cells not part of the phenotypes of interest will be coloured "lightgrey"
}
