% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_double_rings.R
\name{simulate_double_rings}
\alias{simulate_double_rings}
\title{Simulate double immune rings}
\usage{
simulate_double_rings(
  background_sample = bg1,
  bg_type = "Others",
  n_double_rings = 2,
  win = NULL,
  properties_of_double_rings = list(I1 = list(name_of_cluster_cell = "Tumour", size =
    300, shape = "Circle", centre_loc = data.frame(x = 1000, y = 1000),
    infiltration_types = c("Immune1", "Immune2", "Others"), infiltration_proportions =
    c(0.15, 0.05, 0.05), name_of_ring_cell = "Immune1", immune_ring_width = 150,
    immune_ring_infiltration_types = c("Others"), immune_ring_infiltration_proportions =
    c(0.15), name_of_double_ring_cell = "Immune2", double_ring_width = 100,
    double_ring_infiltration_types = c("Others"), double_ring_infiltration_proportions =
    c(0.15)),      I2 = list(name_of_cluster_cell = "Tumour", size = 300, shape = "Oval",
    centre_loc = data.frame(x = 1200, y = 1200), infiltration_types = c("Immune1",
    "Immune2", "Others"), infiltration_proportions = c(0.15, 0.05, 0.05),
    name_of_ring_cell = "Immune1", immune_ring_width = 150,
    immune_ring_infiltration_types = c("Others"), immune_ring_infiltration_proportions =
    c(0.15), name_of_double_ring_cell = "Immune2", double_ring_width = 100,
    double_ring_infiltration_types = c("Others"), double_ring_infiltration_proportions =
    c(0.15))),
  plot.image = TRUE
)
}
\arguments{
\item{background_sample}{(OPTIONAL) Data.Frame The image that the patterns
are simulated on. By default use the internal \code{bg1} background image.}

\item{bg_type}{(OPTIONAL) String The name of the background cell type. By
default is "Others".}

\item{n_double_rings}{Number of double immune rings. This must match the
\code{length(properties_of_double_rings)}.}

\item{win}{(OPTIONAL) owin object output from spatstat.geom::owin function.
By default is the window of the background image.}

\item{properties_of_double_rings}{List of properties of the double immune
rings.}

\item{plot.image}{Boolean. Whether the simulated image is plotted.}
}
\value{
A data.frame of the simulated image
}
\description{
Based on an existing background image, simulate double rings of
immune cells that surround tumour clusters. The inner ring is the internal
margin of a tumour cluster. The outer ring is the external tumour margin.
The tumour clusters and the double immune rings are simulated at the same
time.
}
\examples{
set.seed(610)
# manually define the properties of the immune ring
properties_of_double_rings <- list(D1 = list(name_of_cluster_cell = "Tumour",size = 300,
shape = "Circle",centre_loc = data.frame("x" = 1000, "y" = 1000),infiltration_types
= c("Immune1", "Immune2", "Others"),infiltration_proportions = c(0.15, 0.05, 0.05),
name_of_ring_cell = "Immune1",immune_ring_width = 150,immune_ring_infiltration_types
= c("Others"),immune_ring_infiltration_proportions = c(0.15),name_of_double_ring_cell
= "Immune2",double_ring_width = 100,double_ring_infiltration_types = c("Others"),
double_ring_infiltration_proportions = c(0.15)),
D2 = list(name_of_cluster_cell = "Tumour",size = 300,shape = "Oval",centre_loc
= data.frame("x" = 1200, "y" = 1200),infiltration_types = c("Immune1", "Immune2", "Others"),
infiltration_proportions = c(0.15, 0.05, 0.05),name_of_ring_cell = "Immune1",
immune_ring_width = 150,immune_ring_infiltration_types = c("Others"),
immune_ring_infiltration_proportions = c(0.15),name_of_double_ring_cell = "Immune2",
double_ring_width = 100,double_ring_infiltration_types = c("Others"),
double_ring_infiltration_proportions = c(0.15)))

double_ring_image <- simulate_double_rings(background_sample = bg1,
n_double_rings = 2, properties_of_double_rings = properties_of_double_rings)
# library(SPIAT)
# plot_cell_categories(double_ring_image, c("Tumour","Immune1","Immune2"),
# c("red","darkblue","darkgreen"),"Phenotype")
}
\seealso{
\code{\link{simulate_background_cells}} for all cell simulation,
\code{\link{simulate_mixing}} for mixed background simulation,
\code{\link{simulate_clusters}} for cluster simulation,
\code{\link{simulate_immune_rings}} for single immune ring simulation, and
\code{\link{simulate_stripes}} for vessel simulation.

Other simulate pattern functions: 
\code{\link{simulate_background_cells}()},
\code{\link{simulate_clusters}()},
\code{\link{simulate_immune_rings}()},
\code{\link{simulate_mixing}()},
\code{\link{simulate_stripes}()}
}
\concept{simulate pattern functions}
