% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mixing.R
\name{simulate_mixing}
\alias{simulate_mixing}
\title{simulate_mixing}
\usage{
simulate_mixing(
  background_sample,
  names_of_mixing = c("Tumour", "Immune", "Others"),
  mixing_degree = c(0.2, 0.4, 0.4)
)
}
\arguments{
\item{background_sample}{Data.Frame The image that the stripes are simulated on}

\item{names_of_mixing}{Vector of the mixed cell types}

\item{mixing_degree}{Vector of the proportions of the mixed cell types}
}
\value{
A data.frame of the simulated image
}
\description{
simulate_mixing
}
\examples{
set.seed(610)
mix_background <- simulate_mixing(background_sample = bg1,
names_of_mixing = c("Tumour","Immune", "Others"), mixing_degree = c(0.2, 0.4,  0.4))

# NOT RUN
# library(SPIAT)
# plot_cell_categories(mix_background, categories_of_interest = c("Tumour","Immune"),
# colour_vector = c("red", "darkgreen"), feature_colname = "Phenotype")
}
