% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_images_with_double_rings.R
\name{simulate_multiple_images_with_immune_rings}
\alias{simulate_multiple_images_with_immune_rings}
\title{simulate_multiple_images_with_double_rings}
\usage{
simulate_multiple_images_with_immune_rings(
  background_sample = bg1,
  cluster_size = 200,
  ring_shape = 1,
  infiltration = 0,
  ring_width = seq(50, 100, 10),
  cluster_loc_x = 0,
  cluster_loc_y = 0
)
}
\arguments{
\item{cluster_loc_y}{}
}
\value{

}
\description{
simulate_multiple_images_with_double_rings
}
