% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bg1}
\alias{bg1}
\title{Background cells (4915 cells in a 2000*2000 window)}
\format{
A data frame with 4915 rows (cells) and 3 columns:
\describe{
\item{Cell.X.Position}{X coordinate}
\item{Cell.Y.Position}{Y coordinate}
\item{Cell.Type}{cell type}
}
}
\usage{
bg1
}
\description{
A data frame containing 4915 background cells on 2000 by 2000 size window
}
\keyword{datasets}
