% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sce.R
\name{format_sce}
\alias{format_sce}
\title{format_sce}
\usage{
format_sce(data)
}
\arguments{
\item{data}{Dataframe that will be the colData of the sce object}
}
\value{
An SingleCellExperiment object
}
\description{
Format a dataframe of data into a singlecellexperiment class
where the count assay is empty
}
