% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_background_images.R
\name{multiple_background_images}
\alias{multiple_background_images}
\title{multiple_background_images}
\usage{
multiple_background_images(
  background_sample,
  names_of_cell_types = c("Tumour", "Immune", "Others"),
  proportions_of_cell_types = list(rep(0.1, 9), seq(0, 0.4, 0.05), seq(0.9, 0.5,
    -0.05)),
  plot.image = T
)
}
\arguments{
\item{background_sample}{Data.frame or SingleCellExperiment object with locations
of points representing background cells. Further cell types will be simulated
based on this background sample.}

\item{names_of_cell_types}{Vector of strings Names of the cell types to generate}

\item{proportions_of_cell_types}{List of vectors. Each vector is a sequence of
proportions of the corresponding cell type. The length of the vector is how many
images to generate. All vectors should be of the same length, also equal to
the number of images}

\item{plot.image}{Boolean Whether plot the simulated images or not.Default is TRUE.}
}
\value{
A list of sce objects
}
\description{
multiple_background_images
}
\examples{
names_of_cell_types = c("Tumour","Immune","Others")
prop1 <- rep(0.1,9)
prop2 <- seq(0, 0.4, 0.05)
prop3 <- seq(0.9,0.5,-0.05)
set.seed(610)
bg_image_list <- multiple_background_images(background_sample = bg1, names_of_cell_types=
names_of_cells_types, proportions_of_cell_types = list(prop1, prop2, prop3), plot.image=F)
}
