% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_images_with_immune_rings.R
\name{multiple_images_with_immune_rings}
\alias{multiple_images_with_immune_rings}
\title{multiple_images_with_immune_rings}
\usage{
multiple_images_with_immune_rings(
  background_sample = bg1,
  cluster_size = 200,
  ring_shape = 1,
  infiltration = 0,
  ring_width = seq(50, 100, 10),
  cluster_loc_x = 0,
  cluster_loc_y = 0,
  ring_infiltration = seq(0, 0.2, 0.05),
  plot.image = TRUE
)
}
\arguments{
\item{plot.image}{Boolean Whether plot the simulated images or not.Default is TRUE.}
}
\value{

}
\description{
multiple_images_with_immune_rings
}
