% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_background_cells.R
\name{simulate_background_cells}
\alias{simulate_background_cells}
\title{simulate_background_cells}
\usage{
simulate_background_cells(n_cells, width, height, min_d, oversample = 1.2)
}
\arguments{
\item{n_cells}{Number of cells in the background}

\item{width}{Number The width of the image}

\item{height}{Number The height of the image}

\item{min_d}{Number The minimum distance between two cells}

\item{oversample}{Number The multiplier for oversampling}
}
\value{

}
\description{
simulate_background_cells
}
