% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_clusters.R
\name{simulate_clusters}
\alias{simulate_clusters}
\title{simulate_clusters}
\usage{
simulate_clusters(
  background_sample = bg1,
  n_clusters = 2,
  bg_type = "Others",
  win = NULL,
  properties_of_clusters = list(C1 = list(name_of_cluster_cell = "Tumour", size = 300,
    shape = "Oval", centre_loc = data.frame(x = 500, y = 500), infiltration_types =
    c("Immune1", "Others"), infiltration_proportions = c(0.1, 0.05)), C2 =
    list(name_of_cluster_cell = "Immune1", size = 500, shape = "Irregular", centre_loc =
    data.frame(x = 1500, y = 500), infiltration_types = c("Immune2", "Others"),
    infiltration_proportions = c(0.1, 0.05)))
)
}
\arguments{
\item{background_sample}{(OPTIONAL) Data.Frame The image that the clusters
are simulated on. Default is the internal background image}

\item{n_clusters}{Number of clusters}

\item{bg_type}{(OPTIONAL) String The name of the background cell type. By
default is "Others".}

\item{win}{(OPTIONAL) owin object output from spatstat.geom::owin function. By
default is the window of the background image}

\item{properties_of_clusters}{List of properties of the clusters}
}
\value{
A data.frame of the simulated image
}
\description{
simulate_clusters
}
\examples{
set.seed(610)
cluster_image <- simulate_clusters(background_sample = bg1,
n_clusters = 2, properties_of_clusters= list(C1 = list(
name_of_cluster_cell = "Tumour", size = 300, shape = "Oval", centre_loc =
data.frame("x" = 500, "y" = 500), infiltration_types = c("Immune1", "Others"),
infiltration_proportions = c(0.1, 0.05))))

set.seed(1)
# can also use predefined cluster properties
properties_of_clusters <- C_shape2  # `C_shape2` is defined in the package
cluster_image2 <- simulate_clusters(background_sample = bg1,
n_clusters = 2, properties_of_clusters= properties_of_clusters)
}
