% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_double_rings.R
\name{simulate_double_rings}
\alias{simulate_double_rings}
\title{simulate_double_rings}
\usage{
simulate_double_rings(
  background_sample = bg1,
  bg_type = "Others",
  n_double_rings = 2,
  win = NULL,
  properties_of_double_rings = list(I1 = list(name_of_cluster_cell = "Tumour", size =
    300, shape = "Circle", centre_loc = data.frame(x = 1000, y = 1000),
    infiltration_types = c("Immune1", "Immune2", "Others"), infiltration_proportions =
    c(0.15, 0.05, 0.05), name_of_ring_cell = "Immune1", immune_ring_width = 150,
    immune_ring_infiltration_types = c("Others"), immune_ring_infiltration_proportions =
    c(0.15), name_of_double_ring_cell = "Immune2", double_ring_width = 100,
    double_ring_infiltration_types = c("Others"), double_ring_infiltration_proportions =
    c(0.15)),      I2 = list(name_of_cluster_cell = "Tumour", size = 300, shape = "Oval",
    centre_loc = data.frame(x = 1200, y = 1200), infiltration_types = c("Immune1",
    "Immune2", "Others"), infiltration_proportions = c(0.15, 0.05, 0.05),
    name_of_ring_cell = "Immune1", immune_ring_width = 150,
    immune_ring_infiltration_types = c("Others"), immune_ring_infiltration_proportions =
    c(0.15), name_of_double_ring_cell = "Immune2", double_ring_width = 100,
    double_ring_infiltration_types = c("Others"), double_ring_infiltration_proportions =
    c(0.15)))
)
}
\arguments{
\item{background_sample}{(OPTIONAL) Data.Frame The image that the patterns
are simulated on. By default use the internal \code{bg1} background image}

\item{bg_type}{(OPTIONAL) String The name of the background cell type. By
default is "Others".}

\item{n_double_rings}{Number of double immune rings}

\item{win}{(OPTIONAL) owin object output from spatstat.geom::owin function. By
default is the window of the background image}

\item{properties_of_double_rings}{List of properties of the double immune rings}
}
\value{
A data.frame of the simulated image
}
\description{
simulate_double_rings
}
\examples{
set.seed(610)
# manually define the properties of the immune ring
properties_of_double_rings <- list(D1 = list(name_of_cluster_cell = "Tumour",size = size,
shape = "Circle",centre_loc = data.frame("x" = 1000, "y" = 1000),infiltration_types
= c("Immune1", "Immune2", "Others"),infiltration_proportions = c(0.15, 0.05, 0.05),
name_of_ring_cell = "Immune1",immune_ring_width = 150,immune_ring_infiltration_types
= c("Others"),immune_ring_infiltration_proportions = c(0.15),name_of_double_ring_cell
= "Immune2",double_ring_width = 100,double_ring_infiltration_types = c("Others"),
double_ring_infiltration_proportions = c(0.15)),
D2 = list(name_of_cluster_cell = "Tumour",size = size,shape = "Oval",centre_loc
= data.frame("x" = 1200, "y" = 1200),infiltration_types = c("Immune1", "Immune2", "Others"),
infiltration_proportions = c(0.15, 0.05, 0.05),name_of_ring_cell = "Immune1",
immune_ring_width = 150,immune_ring_infiltration_types = c("Others"),
immune_ring_infiltration_proportions = c(0.15),name_of_double_ring_cell = "Immune2",
double_ring_width = 100,double_ring_infiltration_types = c("Others"),
double_ring_infiltration_proportions = c(0.15)))

double_ring_image <- simulate_double_rings(background_sample = bg1,
n_double_rings = 2, properties_of_double_rings = properties_of_double_rings)
# not run
# library(SPIAT)
# plot_cell_categories(double_ring_image, c("Tumour","Immune1","Immune2"),
# c("red","darkblue","darkgreen"),"Phenotype")

set.seed(610)
# can also use predefined immune rings properties
properties_of_immune_rings2 <- D_shape1  # `D_shape1` is defined in the package
double_ring_image2 <- simulate_immune_rings(background_sample = bg1,
n_immune_rings = 2, properties_of_immune_rings= properties_of_immune_rings2)
# not run
# library(SPIAT)
# plot_cell_categories(immune_ring_image2, c("Tumour","Immune1","Immune2"),
c("red","darkblue","darkgreen"),"Phenotype")
}
