% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_immune_rings.R
\name{simulate_immune_rings}
\alias{simulate_immune_rings}
\title{simulate_immune_rings}
\usage{
simulate_immune_rings(
  background_sample,
  bg_type = "Others",
  n_immune_rings = 2,
  win = NULL,
  properties_of_immune_rings = list(I1 = list(name_of_cluster_cell = "Tumour", size =
    700, shape = "Circle", centre_loc = data.frame(x = 1800, y = 1500),
    infiltration_types = c("Immune1", "Immune2", "Others"), infiltration_proportions =
    c(0.15, 0.05, 0.05), name_of_ring_cell = "Immune1", immune_ring_width = 150,
    immune_ring_infiltration_types = c("Others"), immune_ring_infiltration_proportions =
    c(0.15)), I2 = list(name_of_cluster_cell = "Tumour", size = 700, shape = "Circle",
    centre_loc = data.frame(x = 1800, y = 1500), infiltration_types = c("Immune1",     
    "Immune2", "Others"), infiltration_proportions = c(0.15, 0.05, 0.05),
    name_of_ring_cell = "Immune1", immune_ring_width = 150,
    immune_ring_infiltration_types = c("Others"), immune_ring_infiltration_proportions =
    c(0.15)))
)
}
\arguments{
\item{background_sample}{Data.Frame The image that the stripes are simulated on}

\item{bg_type}{(OPTIONAL) String The name of the background cell type. By
default is "Others".}

\item{n_immune_rings}{Number of immune rings}

\item{win}{(OPTIONAL) owin object output from spatstat::owin function. By
default is the window of the background image}

\item{properties_of_immune_rings}{List of properties of the immune rings}
}
\value{

}
\description{
simulate_immune_rings
}
