% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mixing.R
\name{simulate_mixing}
\alias{simulate_mixing}
\title{simulate_mixing}
\usage{
simulate_mixing(
  background_sample,
  names_of_mixing = c("Tumour", "Immune", "Others"),
  mixing_degree = c(0.2, 0.4, 0.4)
)
}
\arguments{
\item{background_sample}{Data.Frame The image that the stripes are simulated on}

\item{names_of_mixing}{Vector of the mixed cell types}

\item{mixing_degree}{Vector of the proportions of the mixed cell types}
}
\value{
A data.frame of the simulated image
}
\description{
simulate_mixing
}
