% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_stripes.R
\name{simulate_stripes}
\alias{simulate_stripes}
\title{simulate_stripes}
\usage{
simulate_stripes(
  background_sample,
  n_stripe_type,
  win = NULL,
  properties_of_stripes = list(S1 = list(number_of_stripes = 1, name_of_stripe_cell =
    "Others", width_of_stripe = 80, infiltration_types = c("Immune1"),
    infiltration_proportions = c(0.08)), S2 = list(number_of_stripes = 1,
    name_of_stripe_cell = "Others", width_of_stripe = 80, infiltration_types =
    c("Immune1"), infiltration_proportions = c(0.08)))
)
}
\arguments{
\item{background_sample}{Data.Frame The image that the stripes are simulated on}

\item{n_stripe_type}{Number of the types of the stripes}

\item{win}{(OPTIONAL) owin object from spatstat owin method. By default it is
the window of the background image.}

\item{properties_of_stripes}{List of the properties of the stripes}
}
\value{
A data.frame of the simulated image
}
\description{
simulate_stripes
}
