% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIS.R
\name{TIS}
\alias{TIS}
\title{TIS}
\usage{
TIS(
  background_sample = NULL,
  n_cells = NULL,
  width = NULL,
  height = NULL,
  min_d = NULL,
  names_of_bg_cells = NULL,
  proportions_of_bg_cells = NULL,
  n_clusters = NULL,
  properties_of_clusters = NULL,
  n_immune_rings = NULL,
  properties_of_immune_rings = NULL,
  n_double_rings = NULL,
  properties_of_double_rings = NULL,
  n_stripe_type = NULL,
  properties_of_stripes = NULL,
  image_name = NULL
)
}
\arguments{
\item{background_sample}{(OPTIONAL) Data.frame or SingleCellExperiment object
with locations of points representing background cells. If NULL, background cells
will be simulated in this function.}

\item{n_cells}{(OPTIONAL) Number of background cells to simulate. If NULL,
already have a background image.}

\item{width}{(OPTIONAL) Number The width of the image.}

\item{height}{(OPTIONAL) Number The height of the image.}

\item{min_d}{(OPTIONAL) Number The minimum distance between two cells.}

\item{names_of_bg_cells}{(OPTIONAL) Vector The cell types of the background
cells. If NULL, the background cells are of one type.}

\item{proportions_of_bg_cells}{(OPTIONAL) Vector The corresponding proportion
of each cell type in the background cells.}

\item{n_clusters}{(OPTIONAL) Number of cell clusters. If NULL, no clusters to
simulate.}

\item{properties_of_clusters}{(OPTIONAL) List of parameters to define the
clusters.}

\item{n_immune_rings}{(OPTIONAL) Number of immune rings. If NULL, no immune
rings to simulate.}

\item{properties_of_immune_rings}{(OPTIONAL) List of parameters to define the
immune rings.}

\item{n_stripe_type}{(OPTIONAL) Number of stripe (vessel) types. If NULL, no
stripes to simulate.}

\item{properties_of_stripes}{(OPTIONAL) List of parameters to define the stripes.}

\item{image_name}{(OPTIONAL) String to name the output tissue image.}
}
\value{
An sce object of the simulated image
}
\description{
Tissue Image Simulator (TIS) integrates the basic simulation functions
in spaSim, including simulating (mixed) background image, clusters, immune rings,
double immune rings and stripes. The patterns are simulated on separate layers
sequentially (e.g. immune rings are simulated after/on top of background cells).

Pattern properties (e.g. \code{properties_of_clusters}) contain the properties of
a pattern in the format of list where each element is one pattern.
These properties can be manually defined. Or users can use predefined properties
from the package (\code{C_shape1}, \code{C_shape2}, \code{C_Shape3} for clusters; \code{R_shape1},
\code{R_shape2}, \code{R_shape3} for immune rings; \code{D_shape1} for double rings).
Details about the format of the properties see \link{simulate_clusters}
}
\examples{
set.seed(610)
double_ring_image <- TIS(background_sample=bg1, n_double_rings = 2,
properties_of_double_rings = D_shape1)

# NOT RUN
# library(SPIAT)
# plot_cell_categories(double_ring_image, categories_of_interest =
# c("Tumour","Immune1", "Immune2"), colour_vector = c("red", "darkgreen",
# "darkblue"), feature_colname = "Phenotype")
}
